analyzer/src/requirements.txt
pymongo
pandas
python-dotenv
flask
gunicorn
pytest
tiktoken
tokonomics
typing_extensions
flask_cors
pyyaml
pysentimiento
transformers
torch
